+++ examples/basic-example/App.tsx
@@ -6,8 +6,8 @@ import * as ExamplePage from './pages/example';
 export default function App() {
   return (
 		<PilotArea>
-			<PilotRoute path={'/'} component={HomePage.default} default={true}/>
-			<PilotRoute path={'/example'} component={ExamplePage.default} getProps={ExamplePage.getStaticProps}/>
+			<PilotRoute path={'/'} Component={HomePage.default}/>
+			<PilotRoute path={'/example'} Component={ExamplePage.default} getProps={ExamplePage.getStaticProps}/>
 		</PilotArea>
   );
 }